refactor(): move run-command command into utils, es6 import for all files
