+++ lib/utils/run-command.util.ts
@@ -1,13 +1,13 @@
 import { resolve } from 'path';
-const spawn = require('win-spawn');
-import { spinner } from '../../ui';
+const spawn = require('win-spawn'); // TODO: replace with cross-spawn package
+import { spinner } from '../ui';
 
 /**
  * Run command in different env
  * @param  {String} commandName     Name of the command
  * @param  {Array}  commandOptions  Options for task
  */
-module.exports = function runCommand(commandName: any, commandOptions: any) {
+export const runCommand = (commandName: string, commandOptions: string[]) => {
 
     if (commandName === 'gulp') {
         commandName = resolve(process.env.npmRoot + '.bin/gulp');
